+++ src/main/java/phenotypage/model/traitementDonneuse/tableau_donneuse/Tableau_DonneuseService.java
@@ -9,7 +9,7 @@ import java.util.Date;
  */
 public interface Tableau_DonneuseService
 {
-	Tableau_Donneuse createTableauDonneuse(Date date, Produit produit, int quantite, String mode_traitement);
+	Tableau_Donneuse createTableauDonneuse(Date date, Produit produit, float quantite, String mode_traitement);
 
 	Tableau_Donneuse save(Tableau_Donneuse tableau_Donneuse);
 }
